requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
fuzzywuzzy==0.18.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
peewee==3.13.3
PyMySQL==0.10.0
pyTelegramBotAPI==3.7.2
python-Levenshtein==0.12.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
